requirements.txt
biopython
pandas
matplotlib
py3Dmol

:-)